The BATs config is now generated in the run_bats
It is no longer pulled from deployments bosh because needs to get the
current directors UUID and thus is no longer static.

We have also changed the place that defines the BAT_DEPLOYMENT_SPEC
environment variable.  It no longer needs to be saved in the secrets
file and pulled from the deployments bosh repo which should help with
readability.  All other environment variables needed by BATs still
follows the pattern of coming out of the secrets.

[#73410854]