This test is not needed.
It looks like this spec tested two deployments of the same release with
different network configurations which should already be covered by
existing bosh tests. It's possible that the original intent of this test
was to prevent IP collision between separate deployments, but that is
not possible at this time so the test is being removed.

[#106529738](https://www.pivotaltracker.com/story/show/106529738)

Signed-off-by: Beyhan Veli <beyhan.veli@sap.com>