Changed the spec to test on VIP networking for AWS and Openstack.
Previously, it was unclear what this spec covered. It used a method
called 'public_ip' which would point to a VIP if it existed, but would
fall back to a private static IP otherwise. The spec now will SSH to a
machine via the VIP and make sure that its ifconfig corresponds with its
private static IP. Additionally, provided a conditional method on the
stemcell to determine whether it supports VIP networking (only AWS and
Openstack do right now).

[#106771316](https://www.pivotaltracker.com/story/show/106771316)

Signed-off-by: Wagner Camarao <wcamarao@vmware.com>