Introduce a spec for static manual networking
We created a spec that will test the static manual IP connectivity from
the director to the deployed VM. We are using a 'ping' command to
accomplish this connectivity test.

Additionally, we refactored the with_release_stemcell_spec.rb so that
the spec descriptions are easier to follow. Removed some unused instance
variables. Removed a focus left on a test from the previous commit.

[#107224760](https://www.pivotaltracker.com/story/show/107224760)

Signed-off-by: Wagner Camarao <wcamarao@vmware.com>