Isolate `delete_vm_reference` cloudcheck test execution
This change forces a redeploy before executing a `delete_vm_reference`
resolution for cck. Initially the test asked for the presence of two
instances, since it's executed in sequence and the previous test would
delete the vm (as a cck resolution); however, increasing the number of
instances introduced errors on other tests, so the corrective measure
we're introducing is to revert to use only one instance and instead
redeploy the test manifest before executing the `delete_vm_reference`
resolution test.

[#164503063](https://www.pivotaltracker.com/story/show/167978643)