Revert "Revert disabling cck_spec on centos"
- talked to systems
- we're reverting eada0c5ef5448c9e80c27aec5e3699a50d2cf891 as the
command is just wrong but that seems like an earlier attempt to fix
centos.
- For now, we just want to skip centos so that their pipeline can be
unblocked. When we want to spend more time verifying cck and centos
compatability we can come back to it, but the story dragged on long
enough that we just want to proceed with most stemcells/cpi's being
happy.

This reverts commit 8c5512038c78557ab927f3fe91c02cf57f954104.