🐞 Accommodate slow Agent startup on vSphere
Thirty seconds wasn't long enough on vSphere to wait for the Agent to
start; on our test run it took 98 seconds, so we increase the timeout
to at least twice that (200 seconds).

We believe the failure was triggered by the removal of the PowerDNS
release, the last big change before the error began appearing. We
speculate that DNS timeouts are causing the slow Agent startup, and that
the presence of PowerDNS reduced that (possibly via negative caching).

A better fix would be configure the vSphere CPI's testbeds with a DNS
server on the jumpbox, but that change is not trivial, and this change
is.

Fixes, when running BATS under vSphere:

```
BATs example failed  'service configuration agent when initially started after instance boot creates a symlink from /etc/sv/monit to /etc/service/monit' ./spec/system/service_configuration_spec.rb:309
```