Fix flakey way of getting agent config
Fixes errors like:
```
Failure/Error: agent_settings = JSON.load(agent_json)

     JSON::ParserError:
       unexpected token at '{"Platform": {"Linux": {"PartitionerType":
       "parted","DevicePathResolutionType": "scsi"}},"Infrastructure":
       {"Settings": {"Sources": [{"Type": "CDROM","FileName":
       "env"}]}}Connection to 192.168.128.121 closed.}'
```

that are a result of manually trying to extract the stdout of a `bosh
ssh` command.  As it turns out, we had an example already of using the
`--result` and `--column` flags to `bosh ssh` to grab only what we need,
and not interleave with stderr.