Attempts to handle scan-and-fix tasks better
This commit attempts to make the BATS tests more robust in the face of
IAAS slowness that causes the Director to take longer than expected to
run the 'scan and fix' task that is spawned after we restart the Agent
on a victim VM.

NOTE: We don't have a retry limit on the "Is there only a scan and fix
task running?" code path, so this COULD never stop waiting.

Let's hope that doesn't happen!

Signed-off-by: Kenneth Lakin <kenneth.lakin@broadcom.com>